IRC6 2007 rearranged few functions to new file irc6_2007.py
